Ignore `DOCKER_CONTENT_TRUST` and print warnings
This is intended to be the first phase of an actual DCT deprecation.

This stops the CLI from reading the `DOCKER_CONTENT_TRUST` environment variables, and prints a warning if it's set at all, if DCT is enabled in some other way (`--disable-content-trust=false`), and if any `docker trust` subcommand is invoked.

Signed-off-by: Tianon Gravi <admwiggin@gmail.com>